+++ src/main/java/org/betonquest/betonquest/quest/event/item/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation for item related events.
+ */
+package org.betonquest.betonquest.quest.event.item;